move prefetchPayload to avoid circular dependency
